Only require galileo if it is enabled and being used
Summary:
There are problems with using galileo with jest
in a test environment. In order to fix those, we should ensure
that galileo is only required if it is going to be used. This
allows for disabling galileo in tests and removing all side
effects caused by loading the galileo client and its dependency
tree.

Reviewers: #web_platform

Differential Revision: https://code.uberinternal.com/D1486159