Make i18n dependency optional
Summary: This is an optional dependency of the Heatpipe emitter just used to get the locale, and so should be an optional dependency of the plugin as well. Having this be required means that the i18n and Rosetta plugins become required dependencies of the default scaffold which is unfortunate because internal apps will normally not need localization.

Test Plan: Tested locally using default scaffold.

Reviewers: ganemone, dennis.lin

Reviewed By: ganemone

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1518759