Make locale negotiation strategy configurable
Summary:
This change introduces the concept of a locale negotiation strategy. This is an
optional dependency that can be used to change the logic around how the
translation locale is determined. It is needed for websites where SEO is a
concern, to allow encoding locales into the URL which will allow crawlers to
discover and index translated pages.

Reviewers: ganemone, #content_platform_— engineering_team, kevhuang

Reviewed By: ganemone, #content_platform_— engineering_team, kevhuang

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1883055