range engines to beginning of the year
Summary:
Motivation: it's disruptive to have to update engines fields every time a new LTS gets released for Node/NPM/yarn. Every time this happens, it interrupts a developer, and support time is needed to fix the engines field. Nonetheless, we want better engines fields than uber-web's `node: >6` because we rely on Node 8 features and Yarn for scaffolding and templates alike.

Yarn updates every few weeks, so pinning to current version is bound to start preventing usage of the tool as early as two weeks from now (and possibly even next week or this week). Instead, allow any yarn version from beginning of year up to now, and only deal with tool usage blockage if a new yarn release ever breaks the world (which should be extremely rare). A `>current` range is as arbitrary as a `>beginning-of-year` range, and both work equally well, so we opt for the least-likely to disrupt choice.

For Node, we only want to prevent users from using old and bleeding-edge versions, but all engines versions this year up to now do work. Using LTS as support target is still the recommendation if a version mismatch somehow causes version-specific problems. Since those types of problems are extremely rare (compared to disruptions due to engines field mismatch) and there's a workflow to solve them, we opt for the least-likely to disrupt engines field version range.

The choice of using NPM 5.6.0 has to do with instability of package lock heuristics in prior versions. This choice also pertains to the chosen Node version cap (8.9.4 is the earliest release that ships w/ NPM 5.6.0).

Reviewers: keving

Reviewed By: keving

Subscribers: keving, jenkins

Differential Revision: https://code.uberinternal.com/D1965355