Allow passing config as a function of ctx
Summary:
It should be possible to change the tealium config based on the fusion context.
For instance, we might change the geo profile based on GeoIP detection. This
change makes it possible to pass config as a function that takes the fusion
context and returns the tealium config, possibly async via a Promise.

Reviewers: #content_platform_— engineering_team, dennis.lin

Reviewed By: dennis.lin

Subscribers: dennis.lin

Maniphest Tasks: T1987475

Differential Revision: https://code.uberinternal.com/D1997681