Add route parameters and error handling to data dependency codemod
Summary:
In bedrock it was common to get the arguments for server side
rendering from the matched react router params. This adds support for that
as well as catching errors thrown by the createRPCPromise thunk. We need
to catch these errors because they are actually thrown as actions, which breaks
server side rendering and doesn't provide any useful error information. Applications
should be able to handle the errors in the failure rpc reducers.

Reviewers: #web_platform, keving

Reviewed By: #web_platform, keving

Subscribers: keving, jenkins

Differential Revision: https://code.uberinternal.com/D2071565