Ensure bedrock compat things are imported even if fusion-core is not there
Summary:
Currently, a migration will run to completion but generate runtime errors such as BedrockCompatToken is not defined. This is because the codemod attempts a relative insertion, but silently fails if an import for `fusion-core` is not found.

This diff ensures that the bedrock imports are added correctly

Other fixes:

- fix tracer registration errors
- fix eslint configuration in monorepo
- fix fusionrc in monorepo
- fix comment placement that would generate linting errors

Reviewers: dennis.lin

Reviewed By: dennis.lin

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2087039