Fix checkYarnRegistry test
Summary:
The test was never actually running since Jest was not waiting for the result of the promise before returning.

See https://github.com/facebook/jest/issues/1377

Fix was to use .resolves instead which will assert that the promise does not reject.

I found this while working on my tests for provisioning integration but thought I'd open a diff here since it's unrelated to what I'm doing.

Reviewers: lhorie, dennis.lin

Reviewed By: lhorie, dennis.lin

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2094597