Improve upgrade command
Summary:
The ./renovate.sh script in this repo is used for greenkeeping both this project and the templates within it. The greenkeeping script, however, is very aggressive and upgrades everything to latest with no regards for potential breaking changes.

This diff adds the ability to make upgrading less aggressive, so that it backs out of an upgrade if tests fail after a dep is upgraded.

The renovate script still runs the aggressive version for now by default so we can continue enforcing that things are in fact green, but the new options allow for more granular control should we need it down the road

This feature is now also a lot more useful to consumers of uber-web itself, as it provides them with a tool that has better batching heuristics than `yarn upgrade`

Reviewers: amsmith, ganemone

Reviewed By: amsmith, ganemone

Subscribers: amsmith, jenkins

Differential Revision: https://code.uberinternal.com/D2219307