fusion-plugin-rosetta: Remove conditional export of tokens
Summary:
Fixes WPT-1807

A previous refactor removed the local type definition of `__NODE__` and moved the definition to a global file. In the process the definition was changed from `Boolean` to `boolean` which is correct. However, `app.register` expects a Token-like object, which `boolean` is not (`Boolean` is probably similar enough as an object type to satisfy Flow).

There are a couple ways to handle this. We could use type coercion to cast the token export to always be the token type so even if the value is false flow still thinks it's the right token type. We could change the signature of app.register to allow bool values. Or we can just remove the conditional export.

This diff proposes the latter option to simplify the exports from the plugin. This keeps the contract consistent where tokens will exist independent of the environment. We lose on tree shaking these out of the browser bundle but token definitions are fairly small in size.

Type coercion would fix this but feels a little dirty to me since we're basically tricking flow into thinking that the exported object will always be available when it actually won't be.

Reviewers: ganemone

Reviewed By: ganemone

Subscribers: ucdbot

Tags: #bugfix_fusionreleaseprocess

Differential Revision: https://code.uberinternal.com/D2320725