Improve dev-cli developer experience
Summary:
This diff attempts to address two parts of the fusion-dev-cli that I think are detrimental to the developer experience.

First, the dev-cli should not ask you if you want to start Cerberus. You have no choice about starting Cerberus. If Cerberus isn't running, the cli won't start your application. So we should just do it automatically.

Secondly, the dev-cli should not automatically open the Cerberus status page. Most developers don't need to see the status page when they start a project. The status page is easily discoverable at studio.uberinternal.com. It becomes particularly annoying when you disconnect and reconnect, resulting in a bunch of studio tabs sitting at the top of your browser.

{F97766203}

A related benefit of this change is that users who start Cerberus in a separate tab will no longer be prompted to start Cerberus by the dev-cli.

Test Plan: Tested with web-eats-v2

Reviewers: bdean, ganemone, rvil

Reviewed By: ganemone, rvil

Subscribers: rvil

Tags: #feature_fusionreleaseprocess

Differential Revision: https://code.uberinternal.com/D2341831