fix hanging from piping request-promise
Summary:
After about 5 requests to the `/magellan/static/js/magellan.js` endpoint, the
fusionjs server becomes unresponsive. All requests start to block for at least 2 minutes
(didn't wait to see if it ever returns).

I'm not sure how to write a test for this, but this seems to be fixed when
using `request` instead of `request-promise`. As mentioned in
https://github.com/request/request-promise#api-in-detail , streaming requests
from `request-promise` is discouraged, so I'm guessing there's some kind of
leak leading to unresponsiveness when piping `request-promise`.

Take note that `request` is already a dependency of fusion-plugin-magellan .

Reviewers: ganemone, keving, lhorie

Reviewed By: ganemone

Subscribers: ucdbot

Tags: #bugfix_fusionreleaseprocess

Differential Revision: https://code.uberinternal.com/D2390265