Add the ability to log debug statements
Summary: I've found it difficult to determine the exact M3 call that is being made (no easy way to tail this). My thought is that we can use a environment variable that when set will use `console.log` to print out the exact arguments being passed to the M3 client.

Reviewers: #web_platform, wua, amsmith

Reviewed By: wua

Subscribers: ucdbot

Tags: #feature_fusionreleaseprocess

Differential Revision: https://code.uberinternal.com/D2430985