Generate sentry DSN locally
Summary: Sentry DSN's can be generated via using the service name now instead of needing to generate one from usentry-mgmt. Change the scaffold so that the correct DSN is generated for a new project and also modify the provisioning script to generate the correct DSN in case someone scaffolded an older version already.

Reviewers: amsmith

Reviewed By: amsmith

Subscribers: amsmith, jenkins

Differential Revision: https://code.uberinternal.com/D2698981