Modify scaffold to use zone based Sentry DSN
Summary:
Healthline teams wants us to move to zone based DSN's instead of hard coding to the same domain. This modifies the scaffold to use a dynamic DSN that is based on the current zone that the app is running in.

I tested this zone based DSN with performance dashboard and I could see errors in Healthline.

Optional code mod can be done but won't be part of this diff.

Reviewers: deacy

Reviewed By: deacy

Subscribers: jenkins, harshit

Differential Revision: https://code.uberinternal.com/D2736041