Fix bug where new healthline DSN was being incorrectly flagged as an invalid DSN
Summary:
The previous diff that changed the sentry DSN to a zone based DSN was being flagged as an incorrect DSN due to the sentry step checking against a whitelist of valid DSN types.

Instead, change to a blacklist where we only bother generating a new DSN if the configuration file was the very old style where the config had placeholder text. In that case, generate a valid DSN for them.

Reviewers: amsmith

Reviewed By: amsmith

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2739683