Pass query string after stripping route prefix [fusion-cli]  (#134)
* Revert "Add experimentalSideEffectsTest config (#302)"

This reverts commit d0e4cd8f792fc0908ac69cd171224bbce2e619db.

* Pass query string after stripping route prefix

When stripping out the route prefix, the only changes strip-prefix made
was making sure the route had a `/` if the prefix was the only part of
the url. This change will make sure that the `/` gets added back if the
url has a prefix followed directly by a query string. Before it could
cause the ssr plugin to throw an error because ctx.path would be null.

meta:sha:941a7b3ff66a1345de82d3da4edc289dcf354d90