Add flowconfig codemod helpers (#451) https://github.com/uber/fusionjs/pull/451
*Note: These codemods aren't being applied yet because there is another codemod that needs to be written to be run alongside this. Also, a lot of canary testing to be done.*

Provide utilities for writing codemods against `.flowconfig` with minimal assumptions. At a minimum we need to be able to remove lines from `.flowconfig`, but we can also allow ensuring a line exists and asserting a certain `flow` version. While not specifically required, adding the supported `flow` version to the `.flowconfig` will de-mystify errors produced by using an old version of `flow-bin`.

`.flowconfig` is a custom format, so the parsing is based off what `flow-bin` considers an invalid `.flowconfig` (verified by monkey-testing):
* Outside of a section (anything above the first section), there can be a single blank line or any number of comment lines, but nothing else.
* Empty lines are allowed anywhere else in the file
* Section fields can contain spaces, but not within the brackets (`[]`)
* Lines can contain spaces before, in, or after the regular expression
* Not every section needs to be included and sections can be empty
* Comments can have spaces before, but must begin with `;`, `#`, or `💩`
* Regular expressions are OCaml regular expressions, therefore we can't try to translate them into JS RegExp
* Only the first line in the `[version]` section is respected

Given this `.flowconfig`:
```
[ignore]
; this is a comment
.*/node_modules/fusion-cli

```
The data structure of the parsed config looks like the following:
```js
{
  ignore: [{
      comment: true, trimmed: '; this is a comment', raw: '; this is a comment'
    }, {
      comment: false, trimmed: '.*/node_modules/fusion-cli', raw: '.*/node_modules/fusion-cli'
    }, {
      comment: false, trimmed: '', raw: ''
  }]
}
```

Whitespace and comments are preserved so as to not lead to a confusing developer experience

Co-authored-by: Mickey Burks <mburks@uber.com>