Apply flow codemods (#480) https://github.com/uber/fusionjs/pull/480
Duplicate of accidentally closed https://github.com/uber/fusionjs/pull/478

* Ensure minimal flow version in `.flowconfig`
* Remove Fusion packages from `.flowconfig` "libs" section
* Run `flow-typed` after bumping versions to install official libdefs

It would be nice to have these versions be dynamic based on what is used in the monorepo, but I think this will suffice for now.

The libdefs that we currently ship are:
```
- fusion-core
  - koa ^2.7.0
  - tape-cup ^4.7.1
- fusion-plugin-i18n
  - locale ^0.1.0
  - tape-cup ^4.7.1
- fusion-plugin-react-redux
  - redux ^4.0.1
  - tape-cup ^4.7.1
- fusion-plugin-redux-action-emitter-enhancer
  - redux ^4.0.1
  - tape-cup ^4.7.1
- fusion-plugin-rpc-redux-react
  - redux-reactors ^1.0.3
  - redux ^4.0.1
```

The scaffold `.flowconfig` looked something like this:
```
[libs]
./node_modules/fusion-plugin-react-redux/flow-typed/npm/redux_v4.x.x.js
./node_modules/fusion-plugin-redux-action-emitter-enhancer/flow-typed/npm/redux_v4.x.x.js
./node_modules/fusion-core/flow-typed/npm/koa_v2.x.x.js
./node_modules/fusion-core/flow-typed/tape-cup_v4.x.x.js
```

Next we'll also be removing the `flow-typed` directories from the `files` field of each of these packages, so we're no longer shipping them.

Co-authored-by: Mickey Burks <mburks@uber.com>