Fix unsupported characters in docker tag name breaking CI (#506) https://github.com/uber/fusionjs/pull/506
The previous PR (https://github.com/uber/fusionjs/pull/501) optimized the cache strategy and this worked. But branch names have support for special characters (like backslashes) that docker tag names do not.

The logic was sound but we need to sanitize the branch name before we utilize it as a tag name. This fix moves the pipeline to be dynamically generated so we can utilize bash string replacement to sanitize the branch name before it's used.

Tested on the public repo as well to make sure all is well. Also this branch name has slashes and if CI passes here the fix works.

Co-authored-by: Derek Ju <derek.ju@gmail.com>