Fix and prevent users from being able to scaffold with invalid service names (#511) https://github.com/uber/fusionjs/pull/511
Infra has a service limitation where you cannot name a service starting with a capital letter and this fails if you try to provision with that name with a cryptic error.

Add some error checking to restrict what kinds of service names can be used and also DRY it up.