Add pull retries to docker pull in CI (#539) https://github.com/uber/fusionjs/pull/539
Occasionally some builds are seeing CI failures when attempting to pull docker images to run a job. I can't seem to track down the exact reason but it could have to do with hitting the ECR limit of 200 layer pulls a second. And given how many layers are being generated (around 44 per image), parallelizing about 20 or so jobs would hit that limit though it is a moving window.

The Buildkite plugin has an exponential backoff retry option and this seems to slightly fix it so I'm going to try this out to see if this has any positive effect before exploring other alternatives.