Switch to TerraBlob for asset-proxying (#561) https://github.com/uber/fusionjs/pull/561
This pull request is the first required change to proxy assets via TerraBlob instead of directly via S3 in the fusion-plugin-s3-asset-proxying.

`upload.js` has been changed to upload to both S3 and TerraBlob (required for the initial migration as old deploys can only fetch from S3).
It does this by setting up two S3 clients, one configured for S3 and one for the S3 API of the TerraBlob Gateway. It then uploads to both destinations – now using the Promise API to wait for both uploads.
The S3 upload is kept as is, while the TerraBlob upload is now uploading files into `/prod/web-platform/fusion-upload-assets/<svc>/<assets.js>`.
For tests and for any service that might have used the custom option `endpoint`, we disable uploading to TerraBlob. It would be useful to have stats on whether anyone is using this option or if it’s purely there for the test suite – this could potentially be added in a later change.

`server.js` contains the application code for download assets from S3/TerraBlob and serving them. It only needs to download the asset once, so unless the `endpoint` option has been configured, we’ll download from TerraBlob.