Write out configs while provisioning to allow easy reprovision (#617) https://github.com/uber/fusionjs/pull/617
The provisioning flow still can fail, sometimes due to circumstances we can control and sometimes due to errors it cannot handle just yet. This forces the user to retry the flow by retyping `uber-web provision` and having them go through the 5+ step flow to answer questions which is annoying.

As a stop gap solution to hardening the service so that it doesn't fail so often, this PR adds config writing to a temp directory using the service name as the key. If the user has to reprovision, the CLI will detect that a previous config was written out and allow the user to just reuse that config instead of having to go through the flow all over again.