Fix flow type issues with integration tests (#640) https://github.com/uber/fusionjs/pull/640
Opened a PR in the flow-typed repo (https://github.com/flow-typed/flow-typed/pull/3611) to export all of the types from the puppeteer libdefs. This PR copies over the changes from that PR.

This fixes a flow issue with our scaffold where the `Page` type is not available since it's currently not available.