Allow sanitizing an empty string (#643) https://github.com/uber/fusionjs/pull/643
The current sanitization utilities use a falsy check to determine if a given input value has a sanitized symbol field. One example of this being used is by the react-helmet-async plugin where the title is sanitized before it's rendered out to the template.

This doesn't handle empty strings though since the falsy check evaluates incorrectly and thus tries to cast the object type as a string which renders out the `[object Object]` title.

We can get around this problem by setting a default title in the scaffold base templates but this also adds a fix to check if the input value in the `escape` method is really missing the sanitize symbol by checking that it doesn't equal `undefined`.

Co-authored-by: Leo Horie <lhorie@uber.com>