Update update-flipr-bootstrap script (#693) https://github.com/uber/fusionjs/pull/693
The original script uses `boxer` to fetch flipr configurations for a given namespace. Whether `boxer` works or not is tied to whether `update-uber-home.sh` works for a given user. Which is to say this entire process is very flaky.

Instead, use the solution proposed here (https://stack.uberinternal.com/questions/7652/local-flipr-setup-for-my-service/7663#7663) to just curl the frontend port of flipr once cerberus is initiated to download the configs for a given service.

Also increase the timeout value for connecting to cerberus since cerberus is slow.

For context, this PR is a result of UWA requiring flipr to work and the suggested way to bootstrap the uwa configurations uses this script to do so.