Add codemod to inline react HOCs (#709) https://github.com/uber/fusionjs/pull/709
Add codemod to identify HOC imports for 5 super simple packages and redirect the import to a new local implementation of the HOC. After this, we can completely remove the following packages:
- @uber/fusion-plugin-google-analytics-react
- @uber/fusion-plugin-logtron-react
- @uber/fusion-plugin-m3-react
- @uber/fusion-plugin-tealium-react
- fusion-plugin-universal-events-react

This codemod will run before the `replace-package-imports` codemod, which will in turn remove the rest of the imports and delete the plugins from `package.json` - Actually, full disclosure, the plugins probably won't be removed if the user has already run those codemods. In that case, they would have a few unused deps still installed.

Co-authored-by: Mickey Burks <mburks@uber.com>