Fix timestamps for batched heatpipe events (#746) https://github.com/uber/fusionjs/pull/746
Fixes https://jeng.uberinternal.com/browse/WPT-4697

For `custom-hp-web-events` emitted from the browser, the heatpipe events all have the same timestamp due to the default batching strategy of sending all events to the `_events` server side endpoint every 5 seconds. The timestamp being used is calculated when the server event is generated which is `Date.now()` for all batched events.

The events were already being augmented with a `webEventsMeta` field that has the timestamp when the event was generated on the client so instead of using the server timestamp, use this client side timestamp instead.

Co-authored-by: Chris Deacy <deacy@uber.com>