Disable cerberus detection logic when starting fusion (#748) https://github.com/uber/fusionjs/pull/748
This PR removes the spawn code for Cerberus and replaces it with a simple console.log that tells the user Cerberus isn't running and that network calls may fail.

During the Fusion dev XP exercise, we identified that we should remove starting up Cerberus when a user starts up the scaffold since this results in the user having to wait for cerberus to connect first before fusion-cli even starts to compile.

While we could do this in parallel (compile code and start cerberus), our detection logic has gotten very flaky of late and it doesn't work for me > 50% of the time and users in WPT chat have reported the same.

While we could fix the flaky logic, ultimately in my opinion we should just get out of this business altogether. **I feel like we are trying to be too clever here when users can just run it themselves or come up with their own solutions to start cerberus**. Let's just focus on what we can control instead of trying to guess what will be the best experience for end users.

While there can be improvements we can make to error messaging when network calls fail, I'm putting this PR out first to just remove it.