Add override to upgrade logic for canary upgrades (#799) https://github.com/uber/fusionjs/pull/799
#### Issue
Currently, a canary release of `create-uber-web` will contain the correct canary version set, but will never actually change dependency versions in your `package.json` because canary versions are the equivalent of version `0.0.0` which is always the lower semantic version during comparison.

#### Solution
Override the `semver.gt()` logic if it's a canary version of a fusion package with the "curated" upgrade strategy.

The `strategy` parameter represents version resolution (how we determine the new version) rather than version negotiation (how the new version gets compared to the current version). This solution is not an ideal place for this logic, but there currently isn't a place for this concept.

#### Impact
Any time we publish a canary version set or put a canary version in a scaffold, it will always be used over the current version. This realistically has limited impact.

Co-authored-by: Mickey Burks <mburks@uber.com>