Don't add Bazel target if a related one already exists (#858) https://github.com/uber/fusionjs/pull/858
This change affects how `jazelle install`/`jazelle add` synchronizes BUILD.bazel files after a local dependency is added.

If we setup a storybook as a project, it only needs the source code of the project, rather than its compiled files. So if a storybook project's BUILD.bazel file lists `//a:library` in its `deps` field, there's no need to also add `//a:a`

For regular projects, the BUILD file sync semantics stay the same