Implement scaffold command (#879) https://github.com/uber/fusionjs/pull/879
The `jazelle scaffold` command copies a template project to another folder, edits its name in package.json, BUILD.bazel references, and adds the new project to manifest.json. In addition, you can specify `prescaffold` and `postscaffold` hooks to run arbitrary logic in the scaffolded project.

Note: there's a breaking change, in that the programmatic `scaffold` method was renamed to `init` to match the jazelle CLI command.