Fall back to overrides only if headers are not populated (#1018) https://github.com/uber/fusionjs/pull/1018
Alters the behavior of the `fusion-plugin-auth-headers` plugin to only fall back to the dev overrides if the relevant headers are not populated.

In general, we expect either the headers to be present, or overrides to be supplied.  In development, we do not expect the headers to be present (as requests are not intercepted on one's own machine).  In production we should not be leaking these overrides, and thus the overrides will not be present.

The case where both are present should only occur in testing.  In this case, having the headers take precedence allows for testing in an environment closer to production.  For integration tests in particular, this improves the overall experience.

With this change, we can use Puppeteer to do things like:

```js
  const MOCK_UUID = 'some-uuid;
  const MOCK_USER = 'Alexander Smith';

  const browser = await puppeteer.launch();
  const page = await browser.newPage();
  page.setExtraHTTPHeaders({
    'x-auth-params-user-uuid': MOCK_UUID,
  });

  await page.goto('http://localhost:3000/getUser');
  const element = await page.$('#userName');
  const text = await page.evaluate(element => element.textContent, element);

  expect(text).toEqual(MOCK_USER);

  await page.close();
  await browser.close();
```