Fix scaffolder not updating to latest versions (#1064) https://github.com/uber/fusionjs/pull/1064
The scaffolder would sometimes be considered "unchanged" and thus not published because of two factors:
1. The templates are *copied* rather than being specified as dependencies. This is by design, but the change detection algorithm only takes into hashes of each package and the hashes of its explicit dependencies (recursively).
2. We only create versions at publish time, so when we copy the templates for tarball calculation, they still reference `0.0.0-monorepo`

The end result of the two meant that the scaffold would miss newer versions if the dependencies of the template changed but the template itself was unchanged.

This PR creates a special case for the templates and the scaffolder to handle the implicit "dependency" arising from the copying that occurs.