Accept semver ranges in major version check (#1075) https://github.com/uber/fusionjs/pull/1075
Some projects use a semver range for the node engine instead of a
specific version number. A range is useful for projects like libraries.

This updates the major version validator function to accept either a
specific version or a range string. In the case of a range, it will use
the lowest possible version matching the range and use its major version
to check against the provided `supportedVersion`'s major version.