Don't throw exception from failed i18n loading (#1143) https://github.com/uber/fusionjs/pull/1143
Right now when the load fails the exception gets thrown but "fusion-plugin-i18n-react" does not catch the error resulting in it being sent to the browsers uncaught error handler. With this change instead of having an uncaught error which is harder to debug/monitor a metric will be emitted with the response body if the JSON parse fails (we have been seeing this in our app).