Allow usage of fetch from the server for CSRF enhanced FetchToken (#1156) https://github.com/uber/fusionjs/pull/1156
The server implementation just throws a function that throws an error saying the fetch token is not usable from the server. There is no good reason for this to exist and makes it confusing for end users that FetchToken only works on the client.

Instead, just pass the FetchToken through.