allow extra query parameters to be added to rpc calls (#1184) https://github.com/uber/fusionjs/pull/1184
Allow providing additional query parameters for all RPC calls.

Currently only `localeCode` query param is passed in all RPC calls.

However in some cases there are more parameters like that - for example in web payments we have integration key and other contextual parameters like location or possibly uuid of target entity on behalf of which operation is to be made that need to be shared by all requests.

This PR adds optional token which can be used to provide extra query parameters to be send with all RPC requests.