Add stack traces to logged errors in stdout (#1219) https://github.com/uber/fusionjs/pull/1219
Ensure stack trace gets logged to stdout. I assume this was an oversight when logtron got replaced with stdout/filebeat.

When error objects are stringified, the stack isn't serialized. This is remedied by manually stringifying the stack trace and adding it as top-level property.

I tested this by deploying a patch to example-trips-viewer staging.

Log explorer before:

<img width="997" alt="Screen Shot 2021-03-08 at 2 02 20 PM" src="https://user-images.githubusercontent.com/780408/110416435-945c2700-8048-11eb-858d-18cd5f128088.png">

Log explorer after:

<img width="1039" alt="Screen Shot 2021-03-08 at 2 02 28 PM" src="https://user-images.githubusercontent.com/780408/110416445-97571780-8048-11eb-8cf2-2c6ed27e5cb1.png">