Ensure errors during logging get logged to log explorer (#1220) https://github.com/uber/fusionjs/pull/1220
If an uncaught exception occurs inside of a log mapper, the server would crash and the raw error would be logged to stderr (in Peleton) but not visible in uMonitor via log explorer.

Context: https://t3.uberinternal.com/browse/WPT-7169

This ensures that when an error occurs while emitting a log it will be logged to stdout in the correct format to be picked up and sent to log explorer.

Co-authored-by: Ryan Tsao <ryan.j.tsao@gmail.com>