Use a Set when collecting all keys for translation (#1261) https://github.com/uber/fusionjs/pull/1261
When collecting all keys to translate an array is being used. For web-eats-v2 we are getting a lot of duplicate keys so that when `translateKeys` is called it is repeating work.

Testing this locally the number of keys and the time spent doing the translate both drop significantly:
```
translations with keys set took 21ms, # of keys: 656
translations with keys array took 150ms, # of keys: 10433
```

There are other lint fix changes that occurred when I saved in my IDE.